ai-service/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.29.0
langchain>=0.1.16
langchain-google-genai>=1.0.3
langchain-openai>=0.1.3
langchain-community>=0.0.32
langgraph>=0.0.55
google-generativeai>=0.8.3
openai>=1.23.2
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
pydantic>=2.6.4
httpx>=0.27.0
pytest>=8.1.1
pytest-asyncio>=0.23.6
tiktoken>=0.6.0
numpy>=1.26.4
